// frontend/teampilot/src/app/services/api/v1/model/teamDTO.ts
/**
 * TeamPilot.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PlayerDTO } from './playerDTO';


export interface TeamDTO { 
    teamId?: string | null;
    teamName?: string | null;
    avatarUrl?: string | null;
    teamManagerId?: string | null;
    players?: Array<PlayerDTO> | null;
}