// frontend/teampilot/src/app/services/api/v1/model/tournamentMatchDTO.ts
/**
 * TeamPilot.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MatchDTO } from './matchDTO';
import { TournamentDTO } from './tournamentDTO';


export interface TournamentMatchDTO { 
    tournament?: TournamentDTO;
    match?: MatchDTO;
}