// frontend/teampilot/src/app/services/api/v1/model/userWithNullablePlayerPropsDTO.ts
/**
 * TeamPilot.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface UserWithNullablePlayerPropsDTO {
  userId?: string | null;
  userType?: string | null;
  nickname?: string | null;
  firstName?: string | null;
  lastName?: string | null;
  email?: string | null;
  avatarUrl?: string | null;
  bio?: string | null;
  monthlySalary?: number | null;
  teamId?: string | null;
  teamName?: string | null;
}